backend/requirements.txt
azure-functions
pandas==2.3.3
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
